# api_schema/openapi-schema.json
{
    "swagger": "2.0",
    "info": {
        "title": "SMERP API",
        "description": "Auto-generated docs for SMERP API",
        "version": "v1"
    },
    "basePath": "/api/v1",
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "Basic": {
            "type": "basic"
        }
    },
    "security": [
        {
            "Basic": []
        }
    ],
    "paths": {
...
        "/contacts/{id}/delete/": {
            "post": {
                "operationId": "contacts_delete",
                "description": "",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Contact"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Contact"
                        }
                    }
                },
                "tags": [
                    "contacts"
                ]
            },
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "A unique integer value identifying this contact.",
                    "required": true,
                    "type": "integer"
                }
            ]
        },
        "/contacts/{id}/recover/": {
            "post": {
                "operationId": "contacts_recover",
                "description": "",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Contact"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Contact"
                        }
                    }
                },
                "tags": [
                    "contacts"
                ]
            },
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "A unique integer value identifying this contact.",
                    "required": true,
                    "type": "integer"
                }
            ]
        },
...
        "/inventory-locations/{id}/delete/": {
            "post": {
                "operationId": "inventory-locations_delete",
                "description": "",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/InventoryLocation"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/InventoryLocation"
                        }
                    }
                },
                "tags": [
                    "inventory-locations"
                ]
            },
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "A unique integer value identifying this inventory location.",
                    "required": true,
                    "type": "integer"
                }
            ]
        },
        "/inventory-locations/{id}/recover/": {
            "post": {
                "operationId": "inventory-locations_recover",
                "description": "",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/InventoryLocation"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/InventoryLocation"
                        }
                    }
                },
                "tags": [
                    "inventory-locations"
                ]
            },
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "A unique integer value identifying this inventory location.",
                    "required": true,
                    "type": "integer"
                }
            ]
        },
...
        "/materials/{id}/delete/": {
            "post": {
                "operationId": "materials_delete",
                "description": "",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Material"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Material"
                        }
                    }
                },
                "tags": [
                    "materials"
                ]
            },
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "A unique integer value identifying this material.",
                    "required": true,
                    "type": "integer"
                }
            ]
        },
        "/materials/{id}/recover/": {
            "post": {
                "operationId": "materials_recover",
                "description": "",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Material"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Material"
                        }
                    }
                },
                "tags": [
                    "materials"
                ]
            },
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "A unique integer value identifying this material.",
                    "required": true,
                    "type": "integer"
                }
            ]
        },
...
    },
    "definitions": {
        "Contact": {
            "required": [
                "description"
            ],
            "type": "object",
            "properties": {
                "id": {
                    "title": "ID",
                    "type": "integer",
                    "readOnly": true
                },
                "company": {
                    "title": "Firma",
                    "type": "integer",
                    "x-nullable": true
                },
                "name": {
                    "title": "Ad",
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 1,
                    "x-nullable": true
                },
                "last_name": {
                    "title": "Soyad",
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 1,
                    "x-nullable": true
                },
                "gender": {
                    "title": "Cinsiyet",
                    "type": "string",
                    "enum": [
                        "e",
                        "k",
                        "b"
                    ]
                },
                "role": {
                    "title": "Rol",
                    "type": "string",
                    "enum": [
                        "owner",
                        "employee",
                        "salesperson",
                        "procurement"
                    ],
                    "x-nullable": true
                },
                "e_mail": {
                    "title": "E-posta",
                    "type": "string",
                    "maxLength": 254,
                    "minLength": 1,
                    "x-nullable": true
                },
                "phone": {
                    "title": "Telefon",
                    "type": "string",
                    "maxLength": 12,
                    "minLength": 1,
                    "x-nullable": true
                },
                "description": {
                    "title": "Açıklama",
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 1
                },
                "created_by": {
                    "title": "Created by",
                    "type": "string",
                    "readOnly": true
                },
                "created_at": {
                    "title": "Created at",
                    "type": "string",
                    "readOnly": true
                }
            }
        },
...
        "InventoryLocation": {
            "type": "object",
            "properties": {
                "id": {
                    "title": "ID",
                    "type": "integer",
                    "readOnly": true
                },
                "name": {
                    "title": "İsim",
                    "type": "string",
                    "maxLength": 50
                },
                "facility": {
                    "title": "Tesis",
                    "type": "string",
                    "enum": [
                        "esenyurt"
                    ],
                    "default": "esenyurt"
                },
                "area": {
                    "title": "Bölge",
                    "type": "integer",
                    "maximum": 2147483647,
                    "minimum": -2147483648,
                    "x-nullable": true
                },
                "section": {
                    "title": "Sıra",
                    "type": "integer",
                    "maximum": 2147483647,
                    "minimum": -2147483648,
                    "x-nullable": true
                },
                "shelf": {
                    "title": "Raf",
                    "type": "integer",
                    "maximum": 2147483647,
                    "minimum": -2147483648,
                    "x-nullable": true
                },
                "bin": {
                    "title": "Bölme",
                    "type": "integer",
                    "maximum": 2147483647,
                    "minimum": -2147483648,
                    "x-nullable": true
                },
                "type": {
                    "title": "Tipi",
                    "type": "string",
                    "enum": [
                        "industrial_shelf",
                        "exit_transit",
                        "injection_transit",
                        "manufacturing_transit",
                        "resin_storage",
                        "masterbatch",
                        "chemical",
                        "unlabeled",
                        "temporary"
                    ]
                },
                "created_by": {
                    "title": "Created by",
                    "type": "string",
                    "readOnly": true
                },
                "created_at": {
                    "title": "Created at",
                    "type": "string",
                    "readOnly": true
                }
            }
        },
...
        "Material": {
            "required": [
                "description"
            ],
            "type": "object",
            "properties": {
                "id": {
                    "title": "ID",
                    "type": "integer",
                    "readOnly": true
                },
                "name": {
                    "title": "Ad",
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 1,
                    "x-nullable": true
                },
                "category": {
                    "title": "Kategori",
                    "type": "string",
                    "enum": [
                        "supplied",
                        "cardboard",
                        "part",
                        "good",
                        "administrative",
                        "pallet",
                        "undefined"
                    ]
                },
                "internal_code": {
                    "title": "İç Kod",
                    "type": "string",
                    "readOnly": true,
                    "minLength": 1,
                    "x-nullable": true
                },
                "description": {
                    "title": "Açıklama",
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 1
                },
                "created_by": {
                    "title": "Created by",
                    "type": "string",
                    "readOnly": true
                },
                "created_at": {
                    "title": "Created at",
                    "type": "string",
                    "readOnly": true
                }
            }
        },
...
    }
}